+++ src/runtime/icalls/system_appdomain.cpp
@@ -159,7 +159,7 @@ RtResult<vm::RtReflectionAssembly*> SystemAppDomain::load_assembly(vm::RtAppDoma
                                                                    vm::RtStackCrawlMark* stack_crawl_mark)
 {
     utils::StringBuilder name_buf;
-    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(name), vm::String::get_length(name), name_buf);
+    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(name), static_cast<size_t>(vm::String::get_length(name)), name_buf);
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(metadata::RtAssembly*, loaded_ass,
                                             vm::Assembly::load_by_name(this_domain, name_buf.as_cstr(), evidence, ref_only, *stack_crawl_mark));
     return vm::Reflection::get_assembly_reflection_object(loaded_ass);
