+++ src/runtime/icalls/system_reflection_assembly.cpp
@@ -48,7 +48,7 @@ RtResult<vm::RtReflectionType*> SystemReflectionAssembly::internal_get_type(vm::
 {
     (void)module; // unused
     utils::StringBuilder name_buf;
-    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(name), vm::String::get_length(name), name_buf);
+    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(name), static_cast<size_t>(vm::String::get_length(name)), name_buf);
     name_buf.sure_null_terminator_but_not_append();
 
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(
@@ -96,7 +96,7 @@ RtResultVoid SystemReflectionAssembly::internal_get_assembly_name(vm::RtString* 
 RtResult<vm::RtReflectionAssembly*> SystemReflectionAssembly::load_from(vm::RtString* path, bool ref_only, int32_t* mark)
 {
     utils::StringBuilder name_buf;
-    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(path), vm::String::get_length(path), name_buf);
+    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(path), static_cast<size_t>(vm::String::get_length(path)), name_buf);
     vm::RtAppDomain* current_app_domain = vm::AppDomain::get_default_appdomain();
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(
         metadata::RtAssembly*, loaded_ass, vm::Assembly::load_by_name(current_app_domain, name_buf.as_cstr(), nullptr, ref_only, *(vm::RtStackCrawlMark*)mark));
@@ -107,7 +107,7 @@ RtResult<vm::RtReflectionAssembly*> SystemReflectionAssembly::load_file_internal
 {
     (void)mark;
     utils::StringBuilder name_buf;
-    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(path), vm::String::get_length(path), name_buf);
+    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(path), static_cast<size_t>(vm::String::get_length(path)), name_buf);
 
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(metadata::RtAssembly*, loaded_ass, vm::Assembly::load_by_name(name_buf.as_cstr()));
     return vm::Reflection::get_assembly_reflection_object(loaded_ass);
@@ -117,7 +117,7 @@ RtResult<vm::RtReflectionAssembly*> SystemReflectionAssembly::load_with_partial_
 {
     (void)evidence;
     utils::StringBuilder name_buf;
-    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(name), vm::String::get_length(name), name_buf);
+    utils::StringUtil::utf16_to_utf8(vm::String::get_chars_ptr(name), static_cast<size_t>(vm::String::get_length(name)), name_buf);
     name_buf.sure_null_terminator_but_not_append();
 
     DECLARING_AND_UNWRAP_OR_RET_ERR_ON_FAIL(metadata::RtAssembly*, loaded_ass, vm::Assembly::load_by_name(name_buf.as_cstr()));
